ixiaxstream/src/requirements.txt
cloudshell-shell-core
requests
cloudshell_automation_api